Cleanup unnecessary code Use es2017 async/await to simplify code move code around for better readability (ingenico, hue)
