--- modules/oauth-token.js
@@ -1,14 +1,0 @@
-var oauthToken = null;
-
-function setToken(token) {
-  oauthToken=token;
-}
-
-function getToken() {
-  return oauthToken;
-}
-
-module.exports = {
-  get:getToken,
-  set:setToken
-}
