+++ package.json
@@ -7,7 +7,7 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "engines": {
-    "node": "7.10.0"
+    "node": "8.x"
   },
   "author": "Salesforce Team",
   "license": "ISC",
@@ -21,7 +21,6 @@
     "jsonwebtoken": "^7.1.9",
     "form-data": "^2.1.4",
     "request-promise": "^4.1.1",
-    "episode-7": "^1.1.0",
     "connect-sdk-nodejs": ">=2.2.0 <3.0.0"
   }
 }
